Most of the numbers in UPNP spec are unsigned 4 byte integers, so should be u32s
Squashed commit of the following:

commit 1c6f459406ca3f192b2fcf2e3cf537cba9870dfa
Author: carl <900013+carlgreen@users.noreply.github.com>
Date:   Mon May 4 17:56:24 2026 +1200

    clean up some remaining test code that didn't have starting index and requested count as u32

commit 1687212addf0d3f07198fdde235a70d3c861a9b1
Author: carl <900013+carlgreen@users.noreply.github.com>
Date:   Mon May 4 17:54:56 2026 +1200

    requested count is a ui4 (unsigned 4 byte int) which in rust is a u32, take 2

commit ee86b95a2a288a2aebecd6f7c3090458eceb715f
Author: carl <900013+carlgreen@users.noreply.github.com>
Date:   Mon May 4 17:50:21 2026 +1200

    update id is a ui4 (unsigned 4 byte int) which in rust is a u32

commit 8f587f0bf23a07aaa86b28fee4dc64574713d61b
Author: carl <900013+carlgreen@users.noreply.github.com>
Date:   Mon May 4 17:47:04 2026 +1200

    system update id is a ui4 (unsigned 4 byte int) which in rust is a u32

commit e20857c15b0820e781a3ebdbb16e4afe6d679e94
Author: carl <900013+carlgreen@users.noreply.github.com>
Date:   Mon May 4 17:39:40 2026 +1200

    total matches is a ui4 (unsigned 4 byte int) which in rust is a u32

commit 4132d56bdda72744a86e6ebdd37acde9e79d4338
Author: carl <900013+carlgreen@users.noreply.github.com>
Date:   Mon May 4 17:30:54 2026 +1200

    number returned is a ui4 (unsigned 4 byte int) which in rust is a u32

commit a8549289753e7166dd3ac319983647f89fdf5a13
Author: carl <900013+carlgreen@users.noreply.github.com>
Date:   Mon May 4 17:19:30 2026 +1200

    requested count is a ui4 (unsigned 4 byte int) which in rust is a u32

commit 69be3e7abca10c51fbc3907ac9676fc1b700a1ab
Author: carl <900013+carlgreen@users.noreply.github.com>
Date:   Mon May 4 14:42:39 2026 +1200

    starting index is a ui4 (unsigned 4 byte int) which in rust is a u32

+++ src/collection.rs
@@ -122,14 +122,14 @@ impl Track {
 #[derive(Clone, Debug)]
 pub struct Collection {
     last_id: u128,
-    system_update_id: u16, // TODO maintain this value
+    system_update_id: u32, // TODO maintain this value
     pub base: PathBuf,
     artists: Vec<Artist>,
 }
 
 impl Collection {
     #[cfg(test)]
-    pub const fn new(system_update_id: u16, base: PathBuf, artists: Vec<Artist>) -> Self {
+    pub const fn new(system_update_id: u32, base: PathBuf, artists: Vec<Artist>) -> Self {
         Self {
             last_id: 0,
             system_update_id,
@@ -185,7 +185,7 @@ impl Collection {
         collection
     }
 
-    pub const fn get_system_update_id(&self) -> u16 {
+    pub const fn get_system_update_id(&self) -> u32 {
         self.system_update_id
     }
